fix: adapt help-text flag docs to current cli.rs, split test into own file
